+++ lib/app/errors/index.d.ts
@@ -1,0 +1,7 @@
+/**
+ * This probably should be refactored into a standalone package that system can use.
+ *
+ * For now, right here.
+ */
+export declare class NotFoundError extends Error {
+}