+++ src/gql/resolvers/base-resolver.ts
@@ -13,13 +13,16 @@ import {
   Authorized
 } from 'type-graphql';
 import { RumbleshipContext } from '../../app/rumbleship-context';
+import { NodeChangePayload } from '../../app/server/rfi-pub-sub-engine.interface';
+import { NotFoundError } from '../../app/errors';
 import { ClassType } from '../../helpers';
 import {
   Node,
   Connection,
   RelayService,
   NodeNotification,
-  NODE_CHANGE_NOTIFICATION
+  NODE_CHANGE_NOTIFICATION,
+  NodeService
 } from '../relay';
 import { BaseResolverInterface, BaseReadableResolverInterface } from './base-resolver.interface';
 import { RawPayload, createNodeNotification } from './create-node-notification';
@@ -152,10 +155,51 @@ export function createReadOnlyBaseResolver<
     @Subscription(type => notificationClsType, {
       name: `on${capitalizedName}Change`,
       topics: `${NODE_CHANGE_NOTIFICATION}_${capitalizedName}`,
-      nullable: true
+      nullable: true,
+      filter: BaseResolver.filterById
     })
     async onChange(@Root() rawPayload: RawPayload): Promise<NodeNotification<TApi>> {
       return createNodeNotification(rawPayload, this, notificationClsType);
+    }
+
+    static async filterById({
+      payload: rawPayload,
+      args,
+      context
+    }: {
+      payload: RawPayload;
+      args?: { id?: string };
+      context: RumbleshipContext;
+    }): Promise<boolean> {
+      return context.beeline.withAsyncSpan({ name: 'subscription.filter' }, async () => {
+        if (!args?.id) {
+          return true;
+        }
+        const payload: NodeChangePayload = JSON.parse(rawPayload.data.toString());
+        context.beeline.addTraceContext({ 'subscription.filter.id': args.id });
+
+        const oid = new Oid(payload.oid);
+        const { scope } = oid.unwrap();
+        let node;
+        const nodeServices = context.container.get('nodeServices') as object;
+        if (scope in nodeServices) {
+          try {
+            node = await (Reflect.get(nodeServices, scope) as NodeService<Node<object>>).getOne(
+              oid
+            );
+          } catch (error) {
+            context.beeline.addTraceContext(error);
+            if (error instanceof NotFoundError) {
+              context.beeline.addTraceContext({ 'subscription.filter.result': false });
+              return false;
+            }
+            throw error;
+          }
+        }
+        const filtered = node ? node.id.toString() === args?.id : false;
+        context.beeline.addTraceContext({ 'subscription.filter.result': filtered });
+        return filtered;
+      });
     }
   }
   return BaseResolver;